feat: failed thunder graphql

+++ internal/pkg/list/dto/create_list.go
@@ -5,7 +5,7 @@ import "github.com/tommyatchiron/togolist/internal/pkg/list/entities"
 type CreateListInput struct {
 	Title       *string `json:"title" validate:"required" example:"My List"`
 	Description *string `json:"description" validate:"required" example:"This is my list"`
-	Priority    *int    `json:"priority" example:"0"`
+	Priority    *int64  `json:"priority" example:"0"`
 }
 
 func (c *CreateListInput) ToEntity() *entities.List {
